Make the module customizable for newline and indent
